pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smt-index"
version = "0.1.0"
description = "CLI tool for indexing AAS Submodel Templates from IDTA registry and GitHub"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Open SMT Index Contributors" }
]
keywords = ["aas", "submodel-templates", "idta", "industry-4.0", "digital-twin"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.1",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "python-slugify>=8.0.4",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.scripts]
smt-index = "smt_index.cli:app"

[project.urls]
Homepage = "https://github.com/admin-shell-io/open-smt-index"
Repository = "https://github.com/admin-shell-io/open-smt-index"

[tool.hatch.build.targets.wheel]
packages = ["src/smt_index"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
